+++ jobbergate-api/CHANGELOG.md
@@ -3,7 +3,7 @@
 This file keeps track of all notable changes to jobbergate-api
 
 ## Unreleased
-
+- Fixed issue when retrieving large files on get routes after upgrading to FastAPI 0.111
 
 ## 5.2.0a4 -- 2024-06-27
 - Change pydantic.BaseSettings config to use `extra=ignore`